refactor: improoving some points after quick review
