+++ client/src/home/modules/canvas-tools-handlers/WritterEventHandler.js
@@ -1,5 +1,5 @@
 import ToolEventHandler from './parent/ToolEventHandler.js';
-import { getRelativeCursorPos } from '../../../utils/eventUtils.js'
+import { getRelativeCursorPos } from '../../../utils/getRelativeCursorPos.js'
 
 
 // Tentar substituir todas as lógicas que lidam com o contexto ou com outra interface como no drawer
